+++ src/cli/mapshaper-run-command.mjs
@@ -90,7 +90,7 @@ import '../commands/mapshaper-subdivide';
 function commandAcceptsMultipleTargetDatasets(name) {
   return name == 'rotate' || name == 'info' || name == 'proj' ||
     name == 'drop' || name == 'target' || name == 'if' || name == 'elif' ||
-    name == 'else' || name == 'endif';
+    name == 'else' || name == 'endif' || name == 'run';
 }
 
 function commandAcceptsEmptyTarget(name) {